+++ relm4-macros/src/view.rs
@@ -3,20 +3,27 @@ use proc_macro2::Span as Span2;
 use quote::quote;
 use syn::{parse_macro_input, Ident};
 
-use crate::component;
+use crate::token_streams::{TokenStreams, TraitImplDetails};
 use crate::widgets::ViewWidgets;
 
 pub(super) fn generate_tokens(input: TokenStream) -> TokenStream {
     let view_widgets: ViewWidgets = parse_macro_input!(input);
-    let model_name = Ident::new("_", Span2::call_site());
 
-    let component::token_streams::TokenStreams {
+    let TokenStreams {
         error,
         init,
         assign,
         connect,
         ..
-    } = view_widgets.generate_streams(&None, &model_name, None, true);
+    } = view_widgets.generate_streams(
+        &TraitImplDetails {
+            vis: None,
+            model_name: Ident::new("_", Span2::call_site()),
+            sender_name: Ident::new("sender", Span2::call_site()),
+            root_name: None,
+        },
+        true,
+    );
 
     let output = quote! {
         #init