+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
++ macros: Add shorthand syntax for simple input messages
 + core: Improve `SharedState` interface and prefer method names related to `RwLock`
 + macros: Don't generate dead code in the widgets struct
 + macros: Improve error reporting on invalid trait implementations